backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
GitPython==3.1.40
openai==1.3.5
python-dotenv==1.0.0